BUG ISOLATED: Importing `"../connectivity/data_consistency"` was root cause of reducer crash
